# scrape: Optimize protobuf scrape parsing