--- a/scripts/genproto.sh
+++ b/scripts/genproto.sh
@@ -10,27 +10,39 @@ if ! [[ "$0" =~ "scripts/genproto.sh" ]]; then
 	exit 255
 fi
 
-# TODO(bwplotka): Move to buf, this is not OSS agnostic, likely won't work locally.
-if ! [[ $(protoc --version) =~ "3.15.8" ]]; then
-	echo "could not find protoc 3.15.8, is it installed + in PATH? Consider commenting out this check for local flow"
-	exit 255
-fi
+SED=$(which gsed 2>/dev/null || which sed)
 
 # Since we run go install, go mod download, the go.sum will change.
 # Make a backup.
 cp go.sum go.sum.bak
 
-INSTALL_PKGS="golang.org/x/tools/cmd/goimports github.com/gogo/protobuf/protoc-gen-gogofast github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger"
+INSTALL_PKGS="golang.org/x/tools/cmd/goimports github.com/gogo/protobuf/protoc-gen-gogofast github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger github.com/bufbuild/buf/cmd/buf@v1.48.0"
 for pkg in ${INSTALL_PKGS}; do
     GO111MODULE=on go install "$pkg"
 done
+
+# Build client proto via buf.
+pushd ./prompb
+buf generate --path=./io/prometheus/client
+# We have to hack global registry which blocks multiple generated types
+# from the same proto file (the same path+file and full name).
+${SED} -i.bak -E 's/protoimpl.DescBuilder\{/protoimpl.DescBuilder\{FileRegistry:nopFileRegistry{},/g' ./io/prometheus/client/*.pb.go
+${SED} -i.bak -E 's/protoimpl.TypeBuilder\{/protoimpl.TypeBuilder\{TypeRegistry:nopTypeRegistry{},/g' ./io/prometheus/client/*.pb.go
+goimports -w ./io/prometheus/client/*.go
+rm -f ./io/prometheus/client/*.bak
+popd
+
+# TODO(bwplotka): Move write building to buf e.g. once moved out of gogo.
+if ! [[ $(protoc --version) =~ "3.15.8" ]]; then
+	echo "could not find protoc 3.15.8, is it installed + in PATH? Consider commenting out this check for local flow"
+	exit 255
+fi
 
 PROM_ROOT="${PWD}"
 PROM_PATH="${PROM_ROOT}/prompb"
 GOGOPROTO_ROOT="$(GO111MODULE=on go list -mod=readonly -f '{{ .Dir }}' -m github.com/gogo/protobuf)"
 GOGOPROTO_PATH="${GOGOPROTO_ROOT}:${GOGOPROTO_ROOT}/protobuf"
 GRPC_GATEWAY_ROOT="$(GO111MODULE=on go list -mod=readonly -f '{{ .Dir }}' -m github.com/grpc-ecosystem/grpc-gateway)"
-
 DIRS="prompb"
 
 echo "generating code"
@@ -44,14 +56,11 @@ for dir in ${DIRS}; do
 		protoc --gogofast_out=plugins=grpc:. -I=. \
             -I="${GOGOPROTO_PATH}" \
             ./io/prometheus/write/v2/*.proto
-		protoc --gogofast_out=Mgoogle/protobuf/timestamp.proto=github.com/gogo/protobuf/types,paths=source_relative:. -I=. \
-            -I="${GOGOPROTO_PATH}" \
-            ./io/prometheus/client/*.proto
-		sed -i.bak -E 's/import _ \"github.com\/gogo\/protobuf\/gogoproto\"//g' *.pb.go
-		sed -i.bak -E 's/import _ \"google\/protobuf\"//g' *.pb.go
-		sed -i.bak -E 's/\t_ \"google\/protobuf\"//g' *.pb.go
-		sed -i.bak -E 's/golang\/protobuf\/descriptor/gogo\/protobuf\/protoc-gen-gogo\/descriptor/g' *.go
-		sed -i.bak -E 's/golang\/protobuf/gogo\/protobuf/g' *.go
+		${SED} -i.bak -E 's/import _ \"github.com\/gogo\/protobuf\/gogoproto\"//g' *.pb.go
+		${SED} -i.bak -E 's/import _ \"google\/protobuf\"//g' *.pb.go
+		${SED} -i.bak -E 's/\t_ \"google\/protobuf\"//g' *.pb.go
+		${SED} -i.bak -E 's/golang\/protobuf\/descriptor/gogo\/protobuf\/protoc-gen-gogo\/descriptor/g' *.go
+		${SED} -i.bak -E 's/golang\/protobuf/gogo\/protobuf/g' *.go
 		rm -f -- *.bak
 		goimports -w ./*.go ./io/prometheus/client/*.go
 	popd