Fixes https://github.com/prometheus/prometheus/issues/14668

Depends on https://github.com/prometheus/prometheus/pull/15720

* `prompb: Generate client proto using Opaque API, for parsing only.`. This also removes all uses of gogo for both client_model and prompb generated Go code for Prometheus proto exp format.