@bwplotka 2024-12-30
Just OpaqueAPI (no lazy decoding, no string reuse) results.

```
benchstat -filter="/fmt:PromProto" scrape-loop-v1.txt scrape-loop-v2.txt
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/scrape
cpu: Apple M1 Pro
                                 │ scrape-loop-v1.txt │         scrape-loop-v2.txt         │
                                 │       sec/op       │   sec/op     vs base               │
ScrapeLoopAppend/fmt=PromProto-2          54.48µ ± 8%   66.39µ ± 2%  +21.88% (p=0.002 n=6)

                                 │ scrape-loop-v1.txt │         scrape-loop-v2.txt          │
                                 │        B/op        │     B/op      vs base               │
ScrapeLoopAppend/fmt=PromProto-2         75.58Ki ± 1%   57.03Ki ± 3%  -24.55% (p=0.002 n=6)

                                 │ scrape-loop-v1.txt │         scrape-loop-v2.txt         │
                                 │     allocs/op      │  allocs/op   vs base               │
ScrapeLoopAppend/fmt=PromProto-2           810.0 ± 0%   1110.0 ± 0%  +37.04% (p=0.002 n=6)

```