@bwplotka 2024-12-30
Just OpaqueAPI (no lazy decoding, no string reuse) results.

```
benchstat -filter="/fmt:PromProto" scrape-loop-v1.txt scrape-loop-v3.txt
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/scrape
cpu: Apple M1 Pro
                                 │ scrape-loop-v1.txt │         scrape-loop-v3.txt         │
                                 │       sec/op       │   sec/op     vs base               │
ScrapeLoopAppend/fmt=PromProto-2          54.48µ ± 8%   68.65µ ± 4%  +26.02% (p=0.002 n=6)

                                 │ scrape-loop-v1.txt │         scrape-loop-v3.txt          │
                                 │        B/op        │     B/op      vs base               │
ScrapeLoopAppend/fmt=PromProto-2         75.58Ki ± 1%   58.47Ki ± 3%  -22.64% (p=0.002 n=6)

                                 │ scrape-loop-v1.txt │         scrape-loop-v3.txt         │
                                 │     allocs/op      │  allocs/op   vs base               │
ScrapeLoopAppend/fmt=PromProto-2           810.0 ± 0%   1211.0 ± 0%  +49.51% (p=0.002 n=6)
```

Getting worse so far (: