@bwplotka 2024-12-30
Ok, I will proceed with the experiments further, but this path of hacking registries is dangerous and difficult. You pretty much have to have only one generated Go code for one proto file, even if some dependency vendors one. e.g. I can't pass custom registry to unmarshallers without changing globals. We got away with this in the past as we used both new proto API and gogo proto, each having separate registries. See https://protobuf.dev/reference/go/faq/#namespace-conflict

Generally all APIs will be [opaque by default, so it's going to be fun for client_golang which returns client_model dto as an API](https://protobuf.dev/reference/go/opaque-migration/#:~:text=Today%2C%20the%20default%20is%20API_OPEN%2C%20but%20the%20upcoming%20Protobuf%20Edition%202024%20will%20change%20the%20default%20to%20API_OPAQUE.) cc @ArthurSens @kakkoyun

I think I will try similar approach but with a separate proto file in the next PR.